requirements.txt
torch
tqdm
numpy
Pillow
faiss-cpu
transformers
ftfy
regex
